requirements.txt
flask==3.0.3
flask-login==0.6.3
authlib==1.3.1
requests==2.32.3
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
